4site/services/video-slideshow-generator/pyproject.toml
[project]
name = "video-slideshow-generator"
version = "0.1.0"
description = "AI-powered video and slideshow generator for GitHub repositories"
authors = [{name = "Project4Site", email = "team@project4site.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "redis==5.0.1",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "moviepy==1.0.3",
    "pillow==10.1.0",
    "opencv-python==4.8.1.78",
    "numpy==1.25.2",
    "matplotlib==3.8.2",
    "seaborn==0.13.0",
    "openai==1.3.7",
    "elevenlabs==0.2.26",
    "pydub==0.25.1",
    "gtts==2.4.0",
    "python-multipart==0.0.6",
    "aiofiles==23.2.1",
    "jinja2==3.1.2",
    "markupsafe==2.1.3",
    "python-json-logger==2.0.7",
    "httpx==0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "ruff==0.1.6",
    "mypy==1.7.1",
    "pre-commit==3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true